Added an option to keep temp files for debugging purposes. Temp ear directory is not being deleted currently, but it's a separate problem.
svn path=/trunk/; revision=37873


Former-commit-id: 62d26220cac67b8e308768d51a145df273e743f2

+++ v3/ejb/ejb-container/src/main/java/org/glassfish/ejb/embedded/DomainXmlTransformer.java
@@ -122,7 +122,10 @@ public class DomainXmlTransformer {
                 _logger.fine("[DomainXmlTransformer] Creating temp domain file: " + out);
             }
 
-            out.deleteOnExit();
+            if (System.getProperty(EJBContainerProviderImpl.KEEP_TEMPORARY_FILES) == null) {
+                out.deleteOnExit();
+            }
+
             fos = new FileOutputStream(out);
             parser = xif.createXMLEventReader(fis);
